reformating and adding calender options
