+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "add-to-calendar-button-react": "^2.4.3",
     "calendar-link": "^2.2.0",
     "chart.js": "^4.4.0",
     "moment": "^2.29.4",
